requirements.txt
langchain
langchain-huggingface
langchain-community
huggingface_hub
langgraph
python-dotenv
pydantic